remove Validation.scala (#736)
